+++ src/sisl/_help.py
@@ -36,32 +36,32 @@ except ImportError:
     from xml.etree.ElementTree import parse as xml_parse
 
 
-def array_fill_repeat(array, size, cls=None):
+def array_fill_repeat(array, size, axis=-1, cls=None):
     """
     This repeats an array along the zeroth axis until it
     has size `size`. Note that initial size of `array` has
     to be an integer part of `size`.
     """
+    array = np.asarray(array, dtype=cls)
     try:
-        reps = size // len(array)
+        reps = size // array.shape[axis]
     except Exception:
-        array = [array]
-        reps = size // len(array)
-    if size % len(array) != 0:
+        # likely a scalar
+        array = np.expand_dims(array, axis=0)
+        reps = size // array.shape[axis]
+    if size % array.shape[axis] != 0:
         # We do not have it correctly formatted (either an integer
         # repeatable part, full, or a single)
         raise ValueError(
             "Repetition of or array is not divisible with actual length. "
             "Hence we cannot create a repeated size."
         )
-    if cls is None:
-        if reps > 1:
-            return np.tile(array, reps)
-        return array
-    else:
-        if reps > 1:
-            return np.tile(np.array(array, dtype=cls), reps)
-        return np.array(array, dtype=cls)
+
+    if reps > 1:
+        tile_reps = list(1 for _ in array.shape)
+        tile_reps[axis] = reps
+        return np.tile(array, tile_reps)
+    return array
 
 
 @set_module("sisl")